XrTelnet addition and OO refactoring
