+++ netmiko/cisco/__init__.py
@@ -2,10 +2,10 @@ from __future__ import unicode_literals
 from netmiko.cisco.cisco_ios import CiscoIosBase, CiscoIosSSH, CiscoIosTelnet
 from netmiko.cisco.cisco_asa_ssh import CiscoAsaSSH
 from netmiko.cisco.cisco_nxos_ssh import CiscoNxosSSH
-from netmiko.cisco.cisco_xr_ssh import CiscoXrSSH
+from netmiko.cisco.cisco_xr import CiscoXrSSH, CiscoXrTelnet
 from netmiko.cisco.cisco_wlc_ssh import CiscoWlcSSH
 from netmiko.cisco.cisco_s300 import CiscoS300SSH
 from netmiko.cisco.cisco_tp_tcce import CiscoTpTcCeSSH
 
 __all__ = ['CiscoIosSSH', 'CiscoIosTelnet', 'CiscoAsaSSH', 'CiscoNxosSSH', 'CiscoXrSSH',
-           'CiscoWlcSSH', 'CiscoS300SSH', 'CiscoTpTcCeSSH', 'CiscoIosBase']
+           'CiscoXrTelnet', 'CiscoWlcSSH', 'CiscoS300SSH', 'CiscoTpTcCeSSH', 'CiscoIosBase']
